test_service/requirements.txt
python-dotenv==1.1.1
requests==2.32.3
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0